Feat: 사칙연산 구현 및 테스트 코드 작성

+++ src/main/java/org/example/io/ConsoleOutput.java
@@ -7,7 +7,7 @@ public class ConsoleOutput implements Output{
     }
 
     @Override
-    public void printCaculatedResult(long result) {
+    public void printCaculatedResult(int result) {
         System.out.println(result);
     }
 }
