+++ src/main/java/org/example/io/Output.java
@@ -3,5 +3,5 @@ package org.example.io;
 public interface Output {
 
     void printAction(String action);
-    void printCaculatedResult(long result);
+    void printCaculatedResult(int result);
 }
